feat: add minimum backoff strategy

+++ Sources/_PowShims/include/_PowShims.h
@@ -1,4 +1,4 @@
 // Trick from swift-numerics to get pow without Foundation
-static inline __attribute__((__always_inline__)) long pow(long x, long y) {
+static inline __attribute__((__always_inline__)) double pow(double x, double y) {
   return __builtin_pow(x, y);
 }
